// pop-web/src/app/shared/openapi/model/measure.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents a measure value with its unit.  Attributes:     value (float): The numerical value of the measure.     unit (str): The unit of the measure.
 */
export interface Measure { 
    value: number;
    unit: string;
}